Address eslint errors in GeoLoader, GeoExporter
