+++ .eslintrc
@@ -32,5 +32,6 @@
     "react/jsx-no-bind": 1,
     "react/sort-comp": 1,
     "prefer-template": 0,
+    "jsx-quotes": [2, "prefer-single"],
   }
 }
