+++ src/components/GeoLoader/GeoLoader.js
@@ -1,4 +1,4 @@
-import React, { PropTypes, Component } from 'react';
+import React from 'react';
 import { Input, Button, Glyphicon, Modal } from 'react-bootstrap';
 
 import g from '../../core/geo';
@@ -14,10 +14,12 @@ Calls callback with resulting geojson
 
 var GeoLoader = React.createClass({
 
-  mixins: [ModalMixin],
   propTypes: {
     onLoadArea: React.PropTypes.func.isRequired,
+    title: React.PropTypes.string,
   },
+
+  mixins: [ModalMixin],
 
   importError: function () {
     this.setState({
@@ -42,7 +44,9 @@ var GeoLoader = React.createClass({
     return (
       <div>
 
-        <Button onClick={this.open} title={this.props.title}><Glyphicon glyph="open-file" /></Button>
+        <Button onClick={this.open} title={this.props.title}>
+          <Glyphicon glyph='open-file' />
+        </Button>
 
         <Modal show={this.state.showModal} onHide={this.close}>
